server/python-asgi/requirements.txt
fastapi==0.61.2
pydantic[dotenv]==1.7.3
uvicorn==0.12.3
firebase-admin==4.5.0
google-cloud-firestore==2.0.2
grpcio==1.34.0
google-cloud-pubsub==2.2.0
google-cloud-bigquery
google-cloud-error-reporting==1.1.0
google-auth==1.24.0
sentry-sdk==0.20.3
httpx==0.16.1
cloudevents==1.2.0
pyjwt==1.7.1
asgiref==3.3.1
